# openapiJSON.json
{
    "openapi":"3.0.1",
    "info":{
       "title":"OpenAPI definition",
       "version":"v0"
    },
    "servers":[
       {
          "url":"http://localhost:8080",
          "description":"Generated server url"
       }
    ],
    "paths":{
       "/gradovi/{name}":{
          "get":{
             "tags":[
                "zupanija-controller"
             ],
             "operationId":"getGrad",
             "parameters":[
                {
                   "name":"name",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "404":{
                   "description":"Not Found",
                   "content":{
                      "*/*":{
                         "schema":{
                            "type":"string"
                         }
                      }
                   }
                },
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/EntityModelGrad"
                         }
                      }
                   }
                }
             }
          },
          "put":{
             "tags":[
                "zupanija-controller"
             ],
             "operationId":"replaceGrad",
             "parameters":[
                {
                   "name":"name",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "requestBody":{
                "content":{
                   "application/json":{
                      "schema":{
                         "$ref":"#/components/schemas/Grad"
                      }
                   }
                },
                "required":true
             },
             "responses":{
                "404":{
                   "description":"Not Found",
                   "content":{
                      "*/*":{
                         "schema":{
                            "type":"string"
                         }
                      }
                   }
                },
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/EntityModelGrad"
                         }
                      }
                   }
                }
             }
          },
          "delete":{
             "tags":[
                "zupanija-controller"
             ],
             "operationId":"deleteGrad",
             "parameters":[
                {
                   "name":"name",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "404":{
                   "description":"Not Found",
                   "content":{
                      "*/*":{
                         "schema":{
                            "type":"string"
                         }
                      }
                   }
                },
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "type":"object"
                         }
                      }
                   }
                }
             }
          }
       },
       "/gradovi":{
          "get":{
             "tags":[
                "zupanija-controller"
             ],
             "operationId":"getGradovi",
             "responses":{
                "404":{
                   "description":"Not Found",
                   "content":{
                      "*/*":{
                         "schema":{
                            "type":"string"
                         }
                      }
                   }
                },
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/CollectionModelEntityModelGrad"
                         }
                      }
                   }
                }
             }
          },
          "post":{
             "tags":[
                "zupanija-controller"
             ],
             "operationId":"newGrad",
             "requestBody":{
                "content":{
                   "application/json":{
                      "schema":{
                         "$ref":"#/components/schemas/Grad"
                      }
                   }
                },
                "required":true
             },
             "responses":{
                "404":{
                   "description":"Not Found",
                   "content":{
                      "*/*":{
                         "schema":{
                            "type":"string"
                         }
                      }
                   }
                },
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "type":"object"
                         }
                      }
                   }
                }
             }
          }
       },
       "/zupanije":{
          "get":{
             "tags":[
                "zupanija-controller"
             ],
             "operationId":"getZupanije",
             "responses":{
                "404":{
                   "description":"Not Found",
                   "content":{
                      "*/*":{
                         "schema":{
                            "type":"string"
                         }
                      }
                   }
                },
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/CollectionModelEntityModelZupanija"
                         }
                      }
                   }
                }
             }
          }
       },
       "/zupanije/{name}":{
          "get":{
             "tags":[
                "zupanija-controller"
             ],
             "operationId":"getZupanija",
             "parameters":[
                {
                   "name":"name",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "404":{
                   "description":"Not Found",
                   "content":{
                      "*/*":{
                         "schema":{
                            "type":"string"
                         }
                      }
                   }
                },
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/EntityModelZupanija"
                         }
                      }
                   }
                }
             }
          }
       },
       "/zupanije/{name}/gradovi":{
          "get":{
             "tags":[
                "zupanija-controller"
             ],
             "operationId":"getGradoviForZupanija",
             "parameters":[
                {
                   "name":"name",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "404":{
                   "description":"Not Found",
                   "content":{
                      "*/*":{
                         "schema":{
                            "type":"string"
                         }
                      }
                   }
                },
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/CollectionModelEntityModelGrad"
                         }
                      }
                   }
                }
             }
          }
       }
    },
    "components":{
       "schemas":{
          "Grad":{
             "type":"object",
             "properties":{
                "nazivG":{
                   "type":"string"
                },
                "povrsinag_km2":{
                   "type":"integer",
                   "format":"int32"
                },
                "broj_stanovnika_grada":{
                   "type":"integer",
                   "format":"int32"
                },
                "naziv_zupanije":{
                   "type":"string"
                }
             }
          },
          "EntityModelGrad":{
             "type":"object",
             "properties":{
                "nazivG":{
                   "type":"string"
                },
                "povrsinag_km2":{
                   "type":"integer",
                   "format":"int32"
                },
                "broj_stanovnika_grada":{
                   "type":"integer",
                   "format":"int32"
                },
                "naziv_zupanije":{
                   "type":"string"
                },
                "links":{
                   "type":"array",
                   "items":{
                      "$ref":"#/components/schemas/Link"
                   }
                }
             }
          },
          "Link":{
             "type":"object",
             "properties":{
                "rel":{
                   "type":"string"
                },
                "href":{
                   "type":"string"
                },
                "hreflang":{
                   "type":"string"
                },
                "media":{
                   "type":"string"
                },
                "title":{
                   "type":"string"
                },
                "type":{
                   "type":"string"
                },
                "deprecation":{
                   "type":"string"
                },
                "profile":{
                   "type":"string"
                },
                "name":{
                   "type":"string"
                }
             }
          },
          "CollectionModelEntityModelZupanija":{
             "type":"object",
             "properties":{
                "links":{
                   "type":"array",
                   "items":{
                      "$ref":"#/components/schemas/Link"
                   }
                },
                "content":{
                   "type":"array",
                   "items":{
                      "$ref":"#/components/schemas/EntityModelZupanija"
                   }
                }
             }
          },
          "EntityModelZupanija":{
             "type":"object",
             "properties":{
                "naziv":{
                   "type":"string"
                },
                "povrsina_km2":{
                   "type":"integer",
                   "format":"int32"
                },
                "broj_stanovnika":{
                   "type":"integer",
                   "format":"int32"
                },
                "sjediste":{
                   "type":"string"
                },
                "broj_gradova":{
                   "type":"integer",
                   "format":"int32"
                },
                "broj_opcina":{
                   "type":"integer",
                   "format":"int32"
                },
                "zupan":{
                   "type":"string"
                },
                "gustoca_stanovnistva_pokm2":{
                   "type":"number",
                   "format":"float"
                },
                "broj_clanova_u_skupstini":{
                   "type":"integer",
                   "format":"int32"
                },
                "granici_sa_morem":{
                   "type":"string"
                },
                "links":{
                   "type":"array",
                   "items":{
                      "$ref":"#/components/schemas/Link"
                   }
                }
             }
          },
          "CollectionModelEntityModelGrad":{
             "type":"object",
             "properties":{
                "links":{
                   "type":"array",
                   "items":{
                      "$ref":"#/components/schemas/Link"
                   }
                },
                "content":{
                   "type":"array",
                   "items":{
                      "$ref":"#/components/schemas/EntityModelGrad"
                   }
                }
             }
          }
       }
    }
 }